frontend/requirements.txt
gradio>=4.24.0
httpx==0.24.1
python-dotenv==1.0.0